CLI: automatically update after adding
This is less annoying than having to always use two commands for the same thing.
It also has the advantage that if the update fails because the pin configuration
is wrong, it won't be added in the first place. (Previously one could add something
with a typo in the repository and it would only fail on the next `update` command.)