[BREAKING] Add versioning support and rename pins.json to sources.json
The versioning support includes checks and an upgrade procedure. This change
in itself however is breaking and must be adressed manually. To proceed, you
need to rename the pins.json to sources.json and also add a `"version": 0` field
to it. The default.nix can be already updated through the newly provided
`upgrade` command, which will also handle the sources.json changes in the future.