npins import-flake: add a test
this tests that a git flake input can be imported, and that Nix can
evaluate the result.

I also tried to write a similar test for the github API, but this seems
to be impossible: github: flake refs can take a custom host as a
parameter, but Nix will reject that host name if it contains a : to
specify a non-default port.