import-flake: detect default branches correctly
Previously, `niv import-flake' would simply assume that the branch
was called `master' if not given, which is not correct; nix flakes
will use the remote's default branch in such cases.

This mimics the behaviour in the flake importer — however, perhaps
it might be better to move it into the Pin struct itself, allowing
branch pins with no explicit branch name given, which then resolve
to the remote's default branch instead?