check all URLs afterwards
This is a fixup of ddb925e70846e41ca74b6038a346d5fbdfecae6b: "check all
URLs beforehand".

The issue with that approach is that git server implementations are the
wild west, and that there is no single and good way to check for the
liveness of a git repo given its URL. Doing the check before all
operations had the crucial downside that any false positives would
needlessly break npins for that repository.

So instead, the check has been moved to after the actual operations, and
is only triggered on the failure path. This still leads to the same
improvement of error messages, while being guaranteed free of false
positives.