feat: use NixHash as internal hash type
Instead of parsing back and forth from `String` in many places, use
`NixHash` as the internal data representation, and parse outputs returned
from various tools early on.

This depends on a [patch to `nix-compat`](https://cl.snix.dev/c/snix/+/30735),
which fixes the deserialization of `NixHash`; without this patch, valid
hashes in JSON fail to parse. This is because `NixHash` without this
patch only works for `&str` but not for `String`. As a result, the
deserialization fails, and it hits an awful serde edge case where due to
`NixHash` being inside of `Option<OptionalUrlsHashes>` which is also
`flatten`ed, this parsing failure silently turns into a `None`.

Additionally, tarballs use a thin `HashWrapper` to avoid a
deserialization failure due to `serde::flatten`.

Co-authored-by: KFears <kfearsoff@gmail.com>